+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-## [UNRELEASED] neptune-experimental 0.2.2
+## [UNRELEASED] neptune-experimental 0.3.0
+
+### Features
+- Added default run name handling ([#18](https://github.com/neptune-ai/neptune-client-experimental/pull/18))
 
 ### Changes
 - Add custom units and descriptions to tqdm progress bars ([#29](https://github.com/neptune-ai/neptune-client-experimental/pull/29))
+
 
 ## neptune-experimental 0.2.1